[install] up install-peco

+++ install/install-peco.sh
@@ -5,6 +5,11 @@ if [[ `uname` == "Darwin" ]]; then
   exit 0
 fi
 
+if [[ `uname` == "Linux" ]]; then
+  echo "install via apt"
+  exit 0
+fi
+
 rm -rf $DOTFILES/local/peco
 ghbin peco/peco # move into pkg
 mv $DOTFILES/pkg/peco $DOTFILES/local/peco
